Enum safe_transmute::Error

pub enum Error {
    Guard(GuardError),
    Unaligned {
        offset: usize,
    },
    InvalidValue,
}

A transmutation error. This type describes possible errors originating from operations in this crate.

Examples

assert_eq!(guarded_transmute_bool_pedantic(&[0x05]),
           Err(Error::InvalidValue));

Variants
Guard(GuardError)

The data does not respect the target type's boundaries.

Unaligned

The given data slice is not properly aligned for the target type. It would have been properly aligned if offset bytes were shifted (discarded) from the front of the slice.

This is currently unused.

Fields of Unaligned

offset: usize

InvalidValue

The data contains an invalid value for the target type.