Function safe_transmute::guarded_transmute_many

pub unsafe fn guarded_transmute_many<T>(bytes: &[u8]) -> Result<&[T], Error>

View a byte slice as a slice of an arbitrary type.

The byte slice must have at least enough bytes to fill a single instance of a type, extraneous data is ignored.

Examples

// Little-endian
assert_eq!(guarded_transmute_many::<u16>(&[0x00, 0x01, 0x00, 0x02])?,
           &[0x0100, 0x0200]);