Function safe_transmute::guarded_transmute_pod_many

pub fn guarded_transmute_pod_many<T: PodTransmutable>(
    bytes: &[u8]
) -> Result<&[T], Error>

Transmute a byte slice into a single instance of a POD.

The byte slice must have exactly enough bytes to fill a single instance of a type.

Examples

// Little-endian
assert_eq!(guarded_transmute_pod_many::<u16>(&[0x00, 0x01, 0x00, 0x02]).unwrap(),
           &[0x0100, 0x0200]);