Module safe_transmute::pod

Plain data object safe transmute

Functions in this module are guarded from out-of-bounds memory access and from unsafe transmutation target types through the use of the PodTransmutable) trait.

However, they are still not entirely safe because the source data may not be correctly aligned for reading and writing a value of the target type. The effects of this range from less performance (e.g. x86) to trapping or address flooring (e.g. ARM), but this is undefined behavior nonetheless.

Traits

PodTransmutable
Type that can be non-unsafely transmuted into

Functions

guarded_transmute_pod_many_pedantic⚠	[
Deprecated
]
View a byte slice as a slice of POD.

guarded_transmute_pod_many_permissive⚠	[
Deprecated
]
View a byte slice as a slice of a POD type.

guarded_transmute_pod_vec_pedantic⚠	[
Deprecated
]
Transform a byte vector into a vector of POD.

guarded_transmute_pod_vec_permissive⚠	[
Deprecated
]
Transform a byte vector into a vector of POD.

transmute_pod⚠
Transmute a byte slice into a single instance of a POD.

transmute_pod_many⚠
Transmute a byte slice into a single instance of a POD.

transmute_pod_pedantic⚠
Transmute a byte slice into a single instance of a POD.

transmute_pod_vec⚠
Transform a byte vector into a vector of POD.