safe_transmute

Module to_bytes

Functions for transmutation from a concrete type to bytes.

Functions

guarded_transmute_to_bytes_pod_manyDeprecated

Transmute a slice of arbitrary types into a slice of their bytes.

transmute_one_to_bytes

Transmute a single instance of a trivially transmutable type into a slice of its bytes.

transmute_one_to_bytes_mut

Transmute a single instance of a trivially transmutable type into a slice of its bytes.

transmute_to_bytes

Transmute a slice of arbitrary types into a slice of their bytes.

transmute_to_bytes_many_unchecked⚠

Transmute a slice of arbitrary types into a slice of their bytes.

transmute_to_bytes_many_unchecked_mut⚠

Transmute a mutable slice of arbitrary types into a mutable slice of their bytes.

transmute_to_bytes_mut

Transmute a mutable slice of a trivially transmutable type into a mutable slice of its bytes.

transmute_to_bytes_unchecked⚠

Transmute a single instance of an arbitrary type into a slice of its bytes.

transmute_to_bytes_unchecked_mut⚠

Transmute a single mutable instance of an arbitrary type into a mutable slice of its bytes.

transmute_to_bytes_vec

Transmute a vector of elements of an arbitrary type into a vector of their bytes, using the same memory buffer as the former.