Function safe_transmute::trivial::guarded_transmute_pod_many_pedantic

pub unsafe fn guarded_transmute_pod_many_pedantic<T: TriviallyTransmutable>(
    bytes: &[u8]
) -> Result<&[T], Error<'_, u8, T>>

👎 Deprecated since 0.11.0:
see trivial::transmute_many() with PedanticGuard for the equivalent behavior

View a byte slice as a slice of a trivially transmutable type.

The byte slice must have at least enough bytes to fill a single instance of a type, and should not have extraneous data.