Trait safe_transmute::PodTransmutable

pub trait PodTransmutable { }

Type that can be non-unsafely transmuted into

In most cases this is a POD class or a trivially copyable class.

Marker trait for guarded_transmute_pod_*() functions.

Nota bene: bools aren't actually non-unsafe to transmute, because they're restricted to being 0 or 1, which means it's UB to transmute an arbitrary byte to into a bool.

Nota bene: if you transmute into a floating-point type you will have a chance to create signaling NaNs (judgement's still out as to whether they're legal).