Function safe_transmute::trivial::guarded_transmute_pod_many_permissive

pub unsafe fn guarded_transmute_pod_many_permissive<T: TriviallyTransmutable>(
    bytes: &[u8]
) -> Result<&[T], Error<'_, u8, T>>

👎 Deprecated since 0.11.0:
see trivial::transmute_many() with PermissiveGuard for the equivalent behavior

View a byte slice as a slice of a trivially transmutable type.

The resulting slice will have as many instances of a type as will fit, rounded down.